requirements.txt
flask==3.0.2
openai==1.12.0
Pillow==10.2.0
python-dotenv==1.0.1
requests==2.31.0
transformers==4.38.2
torch==2.2.1
langgraph==0.0.26 